Make PVC size configurable and enable conditional provisioning
PVCs are divided into 3 categories:

1) Always provisioned. These are shkeeper and mariadb.

2) Only provisioned if needed. These are storing recoverable data, e.g. blockchain, which can be re-downloaded: tron-fullnode, ethereum-fullnode etc. These PVCs are unprovisioned when related crypto or fullnode is set to `enabled: false` in `values.yaml`.

3) Provisioned, until explicitly turned off. These are storing unrecoverable data such as wallets, crypto keys etc: btc, ltc, doge, monero-rpc-wallet, tron-shkeeper, btc-lightning, which can't be recovered (unless a user has a valid backup).

To disable the provisioning of such PVCs for all existing and newcoming cryptos, a new top-level `values.yaml` option `keepUnusedEssentialPVCs` has been added. If set to `false`, and related crypto is `enabled: false`, the PVC will not be provisioned on `helm install`, or will be removed on `helm upgrade`.

If only a specific PVC needs to be removed, without changing the global `keepUnusedEssentialPVCs` option, it could be done by setting related `storage.size` to `0`, e.g.:

```
doge:
  enabled: false
  dogecoind:
    storage:
      size: 0
```